Fix reconnect_interval on remotes with pcs 0.10
When switching to the pcs 0.10 providers we moved from the the general
pcmk_resource provider to the pcmk_remote backend. While doing this we
lost the ability to specify the reconnect_interval.

Tested by deploying tripleo with the following:
  ExtraConfig:
    pacemaker_remote_reconnect_interval: 180

And correctly observing that the remotes have the right
reconnect_interval:
[root@controller-0 ~]# pcs resource config compute-0 |grep reconnect
  Attributes: reconnect_interval=180 server=172.17.1.96

Change-Id: Idb6e4ba4eadbeeb52b959bc688eb5e043f11c155
Closes-Bug: #1905606