Make pcs resource config/show work depending on the installed pcs version
When we added Id8807a644a6d53f89f99b9db988cbbca99abb7ab we were under
the impression that train and centos-7 was not a thing. Since centos-7
pcs does not have the pcs resource config it is likely we will fail to
recognize existing resources.

This can be removed once we stop supporting centos 7, but for now let's
support both transparently.

Tested on both centos-7 env where it was failing and on a centos-8 train
deployment.

Closes-Bug: #1915519

Change-Id: Ie07b37581e844b64063e5e501f09edf4c8ef24e3