Fix spec tests using the deprecated pacemaker_restart_resource? function
Currently we get the following error:

  1) pcmk_common functions pcmk_restart_resource? stonith resource
     Failure/Error: expect(pcmk_restart_resource?('stonith-fence_ipmilan-stonith-fence-1', "cib-resource.xml")).to eq true

       expected: true
            got: false

       (compared using ==)

       Diff:
       @@ -1 +1 @@
       -true
       +false

     # ./spec/unit/puppet/provider/pcmk_common_spec.rb:41:in `block (2 levels) in <top (required)>'

Reason for this is that there are two functions to detect a change in a
pcmk resource:
- pacemaker_restart_resource? -> which parses crm_simulate and was
  necessary due to https://bugzilla.redhat.com/show_bug.cgi?id=1561617
  (Fixed in pacemaker-1.1.18-12)
- pacemaker_restart_resource_ng? -> which is the preferred one and uses
  crm_diff xml output.

The problem in CI is that crm_simulate output slightly changed and so
one test is failing. Since it makes no sense to call
pacemaker_restart_resource? with a working crm_diff (i.e. any pacemaker
> 1.1.18-12), we just make sure we do not run these tests on a working
crm_diff.

We plan to remove pacemaker_restart_resource? fully now that newton (and
el7) is unsupported, so this change is a minimal non-invasive fix for
the time being.

Closes-Bug: #1921118

Change-Id: Ic8a8e127004220bb8b6735b6cf8b4b52c03fa825