Make sure we test for Promoted when looking for constraints
pacemaker 2.1 returns 'Promoted' as opposed to 'Master' in its
with-rsc-role output. Let's test for both of them to keep compatibility
forwards and backwards. See: https://github.com/ClusterLabs/pacemaker/commit/c3b7b13ef
for example.

Tested this on Centos9 and fixed the issue with puppet not detecting
that a constraint was already there.

Change-Id: Ic215719e52ef1cc8a7a99bafe5d7393f583cc7b7