Add fence_watchdog (sbd) stonith agent
This commit adds a new stonith agent "fence_watchdog", allowing
operators to define a stonith device called "watchdog" (name is
hard coded in pacemaker) and eventually add this to the clusters
stonith topology.

Since this is a special dummy stonith device it is exempted from
having a constraint rule preventing it from running on the same
cluster nodes that is in charge of monitoring.

The "watchdog" resource needs the sbd daemon to be enabled and
depends on a valid watchdog timer device - either phisical or
virtual - to perform self-fencing when required.

Depends-on: https://review.opendev.org/c/openstack/puppet-pacemaker/+/866954

Change-Id: Id010a392df0047d53dfab1c21cc78021c8c1aabf
(cherry picked from commit 397fbfd47097260ed75bd34d4e23b8cac335a43f)