Stable-only: Run stable/wallaby job
puppet-pacemaker 1.1.x is supported by TripleO wallaby or train. This
updates the job definition to run wallaby jobs, so that we can ensure
this module works as part of wallaby release contents.

This also replaces git protocol by https protocol to avoid the failure
during pulling the content. The failure is likely caused by the change
in github which prohibited unauthenticated pull over git. This change
was already made in master and stable/1.5.0 by [1].

Finally, unit tests are temporarily made non-voting, so that we can
unblock gate before we implement the appropriate some version pins to
use libraries compatible with old Ruby.

[1] https://review.opendev.org/c/openstack/puppet-pacemaker/+/836767

Depends-on: https://review.opendev.org/c/openstack/tripleo-ci/+/866722
Change-Id: I10cba06de209a15e7e022692f945891496246ff0